refactor code structure and change storage of countries GeoTable
