+++ Project.toml
@@ -14,6 +14,7 @@ GeoTables = "e502b557-6362-48c1-8219-d30d308dcdb0"
 Meshes = "eacbb407-ea5a-433e-ab97-5258b1ca43fa"
 NaturalEarth = "436b0209-26ab-4e65-94a9-6526d86fea76"
 PrecompileTools = "aea7be01-6a6a-4083-8856-8a6e6704d82a"
+ScopedValues = "7e506255-f358-4e82-b7e4-beb19740aa63"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
 Unitful = "1986cc42-f94f-5a68-af5c-568840ba703d"
 
@@ -35,6 +36,7 @@ Meshes = "0.50, 0.51, 0.52, 0.53"
 NaturalEarth = "0.1"
 PlotlyBase = "0.8"
 PrecompileTools = "1"
+ScopedValues = "1.3.0"
 Tables = "1"
 Unitful = "1"
 julia = "1.11"
