+++ src/GeoTablesConversion.jl
@@ -1,19 +1,18 @@
 module GeoTablesConversion
-    using Meshes
-    using Meshes: Geometry, Manifold, CRS, 🌐, Multi, 𝔼, Point, prettyname, printinds, MultiPolygon
+    using CoordRefSystems: CoordRefSystems, LatLon
+    using Meshes: Meshes, Geometry, Multi, Point, coords, close, GeometrySet, Ring, PolyArea
     using CircularArrays: CircularArray
-    using GeoTables
-    using Tables
+    using GeoTables: GeoTables, georef
+    import Tables
     import GeoInterface as GI
-    using CoordRefSystems
-    using CoordRefSystems: Deg, Met
-    using Unitful
-    using Unitful: °
+    using Unitful: Unitful, @u_str
+
+    # Types
+    using ..CountriesBorders: LATLON, CART, POINT_LATLON, POINT_CART, VALID_POINT, RING_LATLON, RING_CART, VALID_RING, POLY_LATLON, POLY_CART, MULTI_LATLON, MULTI_CART, BOX_LATLON, BOX_CART, CountryBorder
+    # Helpers
+    using ..CountriesBorders: to_cart_point, to_latlon_point, floattype, to_raw_coords, cartesian_geometry, latlon_geometry, change_geometry, remove_polyareas!
 
     export CountryBorder, DOMAIN, remove_polyareas!
-    export change_geometry, latlon_geometry, cartesian_geometry, to_cart_point, to_latlon_point, floattype, to_raw_coords
 
-    include("main_type.jl")
-    include("helpers.jl")
     include("conversion_utils.jl")
 end
